chore: rename dirs
