+++ packages/components.ts
@@ -1,9 +1,9 @@
-export { default as EMap } from "./map/Map/index.vue";
-export { default as EWebMap } from "./map/WebMap/index.vue";
-export { default as EMapView } from "./views/MapView/index.vue";
-export { default as ESceneView } from "./views/SceneView/index.vue";
-export { default as EFeatureLayer } from "./layers/FeatureLayer/index.vue";
-export { default as EMapImageLayer } from "./layers/MapImageLayer/index.vue";
-export { default as ETileLayer } from "./layers/TileLayer/index.vue";
-export { default as EWidget } from "./widgets/Widget/index.vue";
-export { default as EDraw } from "./views/Draw/index.vue";
+export { default as EMap } from "./map/map/index.vue";
+export { default as EWebMap } from "./map/web-map/index.vue";
+export { default as EMapView } from "./views/map-view/index.vue";
+export { default as ESceneView } from "./views/scene-view/index.vue";
+export { default as EFeatureLayer } from "./layers/feature-layer/index.vue";
+export { default as EMapImageLayer } from "./layers/map-image-layer/index.vue";
+export { default as ETileLayer } from "./layers/tile-layer/index.vue";
+export { default as EWidget } from "./widgets/widget/index.vue";
+export { default as EDraw } from "./views/draw/index.vue";
