+++ packages/layers/map-image-layer/index.vue
